\name{knee}
\docType{data}
\alias{knee}
\title{
  knee flexion-extension during walking
}
\description{
  rotation of the knee throughout a walking cycle.
}
\usage{
data(knee)
}
\format{
  a data.frame with the following columns:

  \itemize{
    \item{percent}{percent of the cycle}
    \item{degree}{Knee flexion in degrees.}
  }
}
\details{
  Example used in Dierckx (1993)
}
\source{
  Soudan, K., and Dierckx, P. (1979) Calculation of derivatives and
  Fourier coefficients of human motion data, while using spline
  functions, \emph{ACM Transactions on Mathematical Software}, 4:
  290-294.

  Dierckx, Paul (1993), \emph{Curve and Surface Fitting with Splines},
  Springer.
}
%\seealso{\code{\link{eval.penalty}}}
%\examples{}
% docclass is function
\keyword{datasets}

