\name{moisture}
\docType{data}
\alias{moisture}
\title{
  Volumetric moisture content 
}
\description{
  16 measurements of volumetric moisture content 
}
\usage{
data(moisture)
}
\format{
  a data.frame with the following columns:   

  \item{x}{
    property 
  }
  \item{y}{
    measurement 
  }  
  \item{v}{
    constant 1 
  }  
  \item{w}{
    Weight:  constant 1 except for the first 2 and the last
    observations, being 10, 3, 10, respectively.  
  }  
}
\details{
  Dierckx (1993, Table 7.2, pp. 129-131)

}
\source{
  Dierckx, Paul (1993), \emph{Curve and Surface Fitting with Splines},
  Springer.
}
%\seealso{\code{\link{eval.penalty}}}
\examples{

}
% docclass is function
\keyword{datasets}

