\name{cam}
\docType{data}
\alias{cam}
\title{
  cam data used in a Dierckx example 
}
\description{

}
\usage{
data(cam)
}
\format{

 
}
\details{
  Dierckx (1993)

}
\source{
  Dierckx, Paul (1993), \emph{Curve and Surface Fitting with Splines},
  Springer.
}
%\seealso{\code{\link{eval.penalty}}}
\examples{

}
% docclass is function
\keyword{datasets}

